… saleor/account/tests/test_notifications.py
… saleor/checkout/tests/test_checkout_complete.py
… saleor/csv/tests/test_notifications.py
… saleor/giftcard/tests/test_notifications.py
… saleor/graphql/account/tests/mutations/account/test_account_register.py
… saleor/graphql/account/tests/mutations/account/test_account_request_deletion.py
… saleor/graphql/account/tests/mutations/account/test_request_email_change.py
… saleor/graphql/account/tests/mutations/account/test_send_confirmation_email.py
… saleor/graphql/account/tests/mutations/authentication/test_request_password_reset.py
… saleor/graphql/account/tests/mutations/staff/test_customer_create.py
… saleor/graphql/account/tests/mutations/staff/test_staff_create.py
… saleor/graphql/invoice/tests/test_send_invoice_notification.py
… saleor/graphql/order/tests/mutations/test_order_capture.py
… saleor/graphql/order/tests/mutations/test_order_confirm.py
… saleor/order/notifications.py
… saleor/order/tests/test_notifications.py
… saleor/order/tests/test_order.py
… saleor/plugins/admin_email/tests/test_notify_events.py
… saleor/plugins/admin_email/tests/test_plugin.py
… saleor/plugins/base_plugin.py
… saleor/plugins/manager.py
… saleor/plugins/sendgrid/plugin.py
… saleor/plugins/sendgrid/tests/test_plugin.py
… saleor/plugins/user_email/notify_events.py
… saleor/plugins/user_email/plugin.py
… saleor/plugins/user_email/tests/test_notify_events.py
… saleor/plugins/user_email/tests/test_plugin.py
… saleor/plugins/webhook/plugin.py
… saleor/plugins/webhook/tests/test_webhook.py