I want to merge this change because it introduces some performance improvements listed below:
- Reduce the number of DB queries called for mutations like `orderFulfill` or `orderFulfillmentUpdateTracking`
- Change the logic of generating a payload for the webhook `NOTIFY` event or for the Email plugin. The notification payload will now be generated only when there is an active `NOTIFY` webhook event or the email plugin has a defined active event. Previously, we always generated the payload regardless of whether any webhook or email plugin was active.
- Fix OrderEvent.resolve_app resolver. It was generating additional db queries per event.
- Drop additional DB query from Dataloader: MediaByProductVariantIdLoader 

Breaking change for self-hosted users with custom plugin that uses `notify` hook:
Instead of receiving `payload` as a method argument, the callable instance is passed.
```python
def notify(self, event: Union[NotifyEventType, str], payload: dict, previous_value):
```
to
```python
def notify(
         self,
         event: Union[NotifyEventType, str],
         payload_func: Callable[[], dict],
         previous_value,
     ):
    ...
    payload = payload_func()
```

Link to the internal task: https://linear.app/saleor/issue/MERX-628/optimize-orderfulfillmentupdatetracking-mutation

### Results:
#### OrderFulfill
```graphql
mutation fulfill {
  orderFulfill(
    order: "T3JkZXI6MzcyNDliOGItMmNhZC00MmJhLWI5NDktZDAyMDg3MTRhNTk1"
    input: {
      notifyCustomer: true, 
      lines: [
        {
          orderLineId: "T3JkZXJMaW5lOjhkNDUyMWI2LTRlMDEtNDA2ZS1hNDRkLTA5MTlmZjM2YTQ0MA==", 
          stocks: {quantity: 1, warehouse: "V2FyZWhvdXNlOmViNmU4NGM3LWMwOGUtNDVjYS1hNDVhLTFlYmQzOTg1MjliYw=="}}, 
        {
          orderLineId: "T3JkZXJMaW5lOjViNDZiOTJmLTAyNWItNGNhMS1hZjMwLTAyZWI3MmRjYzVlNA==", 
          stocks: {quantity: 1, warehouse: "V2FyZWhvdXNlOmViNmU4NGM3LWMwOGUtNDVjYS1hNDVhLTFlYmQzOTg1MjliYw=="}
        }
      ]
    }
  ) {
    fulfillments {
      id
    }
    errors {
      field
      code
      message
    }
  }
}
```
Main branch without these changes:
```
Queries: 226 in 0.607000s on /graphql/ [PID:7294:ThreadPoolExecutor-0_0]
Response took 1.036024s [PID:7294:ThreadPoolExecutor-0_0]
```

After improvements, when no active `NOTIFY` webhook & Email plugin not subscribed for fulfillment event
```
Queries: 53 in 0.291000s on /graphql/ [PID:7377:ThreadPoolExecutor-0_0]
Response took 0.424378s [PID:7377:ThreadPoolExecutor-0_0]
```

After improvements, when active NOTIFY webhook or email plugin:
```
Queries: 81 in 0.449000s on /graphql/ [PID:7377:ThreadPoolExecutor-0_0]
Response took 0.753008s [PID:7377:ThreadPoolExecutor-0_0]
```
#### orderFulfillmentUpdateTracking
```graphql
mutation {
  orderFulfillmentUpdateTracking(
    id: "RnVsZmlsbG1lbnQ6NDQ3NzQwNg==",
    input:{trackingNumber:"12112", notifyCustomer:true}
  ) {
    fulfillment {
      id
    }
  }
}
```
Main branch without these changes:
```
Queries: 188 in 0.586000s on /graphql/ [PID:16712:ThreadPoolExecutor-0_0]
Response took 0.864708s [PID:16712:ThreadPoolExecutor-0_0]
```

After improvements, when no active `NOTIFY` webhook & Email plugin not subscribed for fulfillment event
```
Queries: 19 in 0.096000s on /graphql/ [PID:16768:ThreadPoolExecutor-0_0]
Response took 0.155588s [PID:16768:ThreadPoolExecutor-0_0]
```
After improvements, when active NOTIFY webhook or email plugin:
```
Queries: 49 in 0.198000s on /graphql/ [PID:16768:ThreadPoolExecutor-0_0]
Response took 0.349881s [PID:16768:ThreadPoolExecutor-0_0]
```



<!-- Please mention all relevant issue numbers. -->

# Impact

- [ ] New migrations
- [ ] New/Updated API fields or mutations
- [ ] Deprecated API fields or mutations
- [ ] Removed API types, fields, or mutations

# Docs

<!-- Docs are stored in a separate repository: https://github.com/saleor/saleor-docs/. -->
<!-- Please provide a link to the PR that updates documentation for your changes. -->
<!-- If changes in docs are not required, please mention that in the description. -->

- [ ] Link to documentation:

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

- [ ] Privileged queries and mutations are either absent or guarded by proper permission checks
- [ ] Database queries are optimized and the number of queries is constant
- [ ] Database migrations are either absent or optimized for zero downtime
- [ ] The changes are covered by test cases
- [ ] All new fields/inputs/mutations have proper labels added (`ADDED_IN_X`, `PREVIEW_FEATURE`, etc.)
- [ ] All migrations have proper dependencies
- [ ] All indexes are added concurrently in migrations
- [ ] All RunSql and RunPython migrations have revert option defined